# imcb_add_buddy is slow, let's not