@sm00th 2018-06-05
Even though bee_user_new() is in an exported header the function itself is not exported for module use, even if it was I don't think that fixing an issue with imcb_add_buddy by hacking around it in bitlbee-discord is the right way to solve this - there are other plugins that can hit this. 

This is a long known issue (#47) and I even had a patch (not sure if I have it anymore and if I do if it would properly apply) that would change bitlbee-discords user handling to hashtables, but that had little effect without changes on bitlbee's side.

Fixing it in bitlbee requires considerable amount of work, but it is not an excuse to fix it elsewhere. 

This patch can be used by people to try and remedy the issue on their own risk, but I see no reason to merge it to mainline, sorry.