@Alcaro 2018-06-05
bee_user_new is right beside bee_user_by_handle, in both header and source code, with no attributes or comments stating only one should be used by plugins. Are you aware of any documentation I'm not?

Fixing it upstream is indeed the correct long-term solution, I forgot that's an option. Did anyone report it?