@sm00th 2018-06-05
> bee_user_new is right beside bee_user_by_handle, in both header and source code, with no attributes or comments stating only one should be used by plugins. Are you aware of any documentation I'm not?

Hm, it is. I have a feeling I ran into some problems using functions that weren't prefixed with G_MODULE_EXPORT, but seeing bee_user_by_handle without it makes me think it might be a false memory.

> Fixing it upstream is indeed the correct long-term solution, I forgot that's an option. Did anyone report it?

Don't see anything like that in [trac](https://bugs.bitlbee.org/query), but @dequis is definitely aware of the problem.