@dequis 2018-06-05
Yeah, fairly high in my list of things i'd love to do if i had time to allocate to bitlbee. Life is kind of a mess lately (moved last month, moving back now).

One idea was adding a hash table to store buddies that is kept in sync with the original linked list to ensure that all API operations still behave the same. This may or may not be feasible given the amount of direct linked list access from plugins, and whether it is read-only or read-write.

Given that this is a big issue and the API boundaries are awkward here (that linked list shouldn't have been exposed in the first place IMO), I'm OK with minor technically-breaking changes that are known to not affect existing plugins (the builtin ones plus the ones linked from the bitlbee wiki).

Of course direct linked list access should be deprecated and removed at some point in the future.

Alternatively, consider more cache-friendly linked list alternatives, something that is backed by a flat preallocated array but otherwise looks like a linked list may be enough to solve performance problems with no api changes (unless, of course, some plugin decides to keep potentially-outdated pointers around - reallocating invalidates every pointer of the list)