@Corosauce 2018-06-06
Alrighty, finally got a performance test run, did a test with latest of master branch vs this pr, strangely enough with or without the pr, doing a full discord join which includes the minecraft discord there has been a whole minute shaved off the connect time, so my leaving and then rejoining the discord might have messed with something, in a good way, not sure whats going on there, either way, heres my data:

```
with PR 160 fix:

Jun 05 22:45:50 <Corosus>	account discord on
Jun 05 22:47:32 <root>	discord - Logging in: Logged in
Jun 05 22:47:32 <root>	discord - Remote host is closing websocket connection
Jun 05 22:47:32 <root>	discord - Performing soft-reconnect

- 102 seconds to process
```

```
without fix (master branch latest):

Jun 05 22:52:41 <Corosus>	account discord on
Jun 05 22:54:25 <root>	discord - Logging in: Logged in
Jun 05 22:54:28 <root>	discord - Remote host is closing websocket connection
Jun 05 22:54:28 <root>	discord - Performing soft-reconnect

- 104 seconds to process
```

After seeing the similar times, I made sure that I properly pulled the source for the PR and master and that the md5 checksum on the .so file was different for each test.

Noticed I was able to actually connect in fine now too, so I dug out the logs from the testing a few days ago:

```
2 days ago, before i left the minecraft discord and rejoined it, with sm00ths code changes since my issue was made:

Jun 03 21:57:57 <Corosus>	account discord on
Jun 03 22:00:37 <root>	discord - Logging in: Logged in
Jun 03 22:00:42 <root>	discord - Remote host is closing websocket connection
Jun 03 22:00:42 <root>	discord - Performing soft-reconnect
Jun 03 22:00:57 <root>	discord - Error: Invalid session, reconnecting
Jun 03 22:00:57 <root>	discord - Signing off..
Jun 03 22:01:00 <root>	discord - Reconnecting in 5 seconds..

- 160 seconds to process
```

O_o