@sm00th 2018-07-16
Closing this one as it is being solved in bitlbee itself.